Use async, fix comment manage paging bug
